Changed algorithm data component to allow access through functions
